tests/e2e: do not run cipher suite tests in coverage tests
